0.1.10

+++ src/yaml_writer/functions/write_database_file.py
@@ -4,7 +4,7 @@ import yaml
 import src.common.processing_vars as var
 import os.path
 #import ruamel.yaml as yaml
-def write_database_file(self,d: dict):
+def write_database_file(self,d:dict, ignore_existing:bool=False):
     yml_path = 'snowflake/data/' + d['DATABASE_NAME_SANS_ENV'] + '/database.yml'
     
     data = self.create_parent_load_data(yml_path)
@@ -18,7 +18,7 @@ def write_database_file(self,d: dict):
     data['COMMENT'] = self.choose_value_string(data, 'COMMENT', d,'COMMENT', var.EMPTY_STRING)
     data['OWNER']=self.choose_value_string(data, 'OWNER', d,'OWNER' , var.EMPTY_STRING)
     
-    if d['TAGS'] != []:
+    if 'TAGS' in d and d['TAGS'] != []:
         if 'TAGS' in data:
             data['TAGS'] = self.choose_list_objects(d['TAGS'], data['TAGS'])
         else:
@@ -26,7 +26,7 @@ def write_database_file(self,d: dict):
     else:
         data['TAGS']=var.EMPTY_STRING
     
-    if d['GRANTS'] != []:
+    if 'GRANTS' in d and d['GRANTS'] != []:
         if 'GRANTS' in data:
             data['GRANTS'] = self.choose_list_objects(d['GRANTS'], data['GRANTS'])
         else:
@@ -41,8 +41,10 @@ def write_database_file(self,d: dict):
     #yaml_string_converted = yaml_string_converted.replace("'"+var.EMPTY_STRING+"'",'')
     #yaml_str = yaml.dump(data, default_flow_style=False,indent=4).replace(r"''", '') #ruamel version
     yaml_string_converted = self.convert_special_characters_back_in_file(yaml_string)
-    with open(yml_path, "w+") as f:
-        f.write(yaml_string_converted)
+    db_file_exists = os.path.isfile(yml_path)
+    if not(db_file_exists and ignore_existing):
+        with open(yml_path, "w+") as f:
+            f.write(yaml_string_converted)
 
     #with Path(yml_path).open("w", encoding="utf-8") as f:
     #    yaml.dump(data,f,default_style=None,default_flow_style=False, sort_keys=False)
